+++ cli/tests/test_diff_command.rs
@@ -47,6 +47,11 @@ fn test_diff_basic() {
     Modified regular file file3 (file1 => file3):
     Modified regular file file4 (file2 => file4):
     [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+      file3
+      file4
+    [EOF]
     ");
 
     let output = work_dir.run_jj(["diff", "--context=0"]);
@@ -390,6 +395,10 @@ fn test_diff_empty() {
     Added regular file file1:
         (empty)
     [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+      file1
+    [EOF]
     ");
 
     work_dir.run_jj(["new"]).success();
@@ -617,6 +626,10 @@ fn test_diff_types() {
     insta::assert_snapshot!(diff("missing", "file"), @r"
     -F foo
     [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+      foo
+    [EOF]
     ");
     insta::assert_snapshot!(diff("file", "conflict"), @r"
     FC foo
@@ -653,6 +666,11 @@ fn test_diff_name_only() {
     deleted
     modified
     [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+      deleted
+      modified
+    [EOF]
     ");
     work_dir.run_jj(["commit", "-mfirst"]).success();
     work_dir.remove_file("deleted");
@@ -665,6 +683,9 @@ fn test_diff_name_only() {
     deleted
     modified
     sub/added
+    [EOF]
+    ------- stderr -------
+    Tracking added, sub/added
     [EOF]
     ");
 }
@@ -690,6 +711,11 @@ fn test_diff_renamed_file_and_dir() {
     insta::assert_snapshot!(output.normalize_backslash(), @r"
     R {y => x}/file
     R {x => y}
+    [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+      x/file
+      y
     [EOF]
     ");
 
@@ -3087,6 +3113,8 @@ fn test_diff_external_tool() {
     insta_portable_exit_status.bind(|| {
         insta::assert_snapshot!(output, @r"
         ------- stderr -------
+        Auto-tracking 1 new file:
+          file3
         Warning: Tool exited with <exit status>: 1 (run with --debug to see the exact invocation)
         [EOF]
         ");
@@ -3301,6 +3329,10 @@ fn test_diff_external_available_width() {
     insta::assert_snapshot!(output, @r"
     50
     [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+      file2
+    [EOF]
     ");
 
     // File-by-file diff
@@ -3371,6 +3403,11 @@ fn test_diff_external_file_by_file_tool() {
     file1
     --
     file4
+    [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+      file3
+      file4
     [EOF]
     ");
 
@@ -3595,6 +3632,10 @@ fn test_diff_stat() {
     file1 | 1 +
     1 file changed, 1 insertion(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+      file1
+    [EOF]
     ");
 
     work_dir.run_jj(["new"]).success();
@@ -3643,11 +3684,19 @@ fn test_diff_stat_long_name_or_stat() {
     一 | 1 +
     2 files changed, 2 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+      1
+      一
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 1, 10), @r"
     1  | 10 ++++++++++
     一 | 10 ++++++++++
     2 files changed, 20 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
     // 30 column display width means right edge is
@@ -3657,11 +3706,17 @@ fn test_diff_stat_long_name_or_stat() {
     一 | 100 +++++++++++++++++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 10, 1), @r"
     1234567890        | 1 +
     ...四五六七八九十 | 1 +
     2 files changed, 2 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 10, 10), @r"
@@ -3669,11 +3724,17 @@ fn test_diff_stat_long_name_or_stat() {
     ...五六七八九十  | 10 ++++++++
     2 files changed, 20 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 10, 100), @r"
     1234567890       | 100 +++++++
     ...五六七八九十  | 100 +++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 50, 1), @r"
@@ -3681,17 +3742,26 @@ fn test_diff_stat_long_name_or_stat() {
     ...四五六七八九十 | 1 +
     2 files changed, 2 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 50, 10), @r"
     ...8901234567890 | 10 ++++++++
     ...五六七八九十  | 10 ++++++++
     2 files changed, 20 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 50, 100), @r"
     ...8901234567890 | 100 +++++++
     ...五六七八九十  | 100 +++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
 
@@ -3703,11 +3773,17 @@ fn test_diff_stat_long_name_or_stat() {
     ...八九十一二三  | 100 +++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 14, 100), @r"
     12345678901234   | 100 +++++++
     ...九十一二三四  | 100 +++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 15, 100), @r"
@@ -3715,11 +3791,17 @@ fn test_diff_stat_long_name_or_stat() {
     ...十一二三四五  | 100 +++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 16, 100), @r"
     1234567890123456 | 100 +++++++
     ...一二三四五六  | 100 +++++++
     2 files changed, 200 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
 
@@ -3731,6 +3813,9 @@ fn test_diff_stat_long_name_or_stat() {
     ... | 10 ++
     2 files changed, 20 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     test_env.add_env_var("COLUMNS", "3");
     let work_dir = test_env.work_dir("repo");
@@ -3739,17 +3824,26 @@ fn test_diff_stat_long_name_or_stat() {
     ... | 10 ++
     2 files changed, 20 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 3, 10), @r"
     123 | 10 ++
     ... | 10 ++
     2 files changed, 20 insertions(+), 0 deletions(-)
     [EOF]
+    ------- stderr -------
+    Tracking 2 files
+    [EOF]
     ");
     insta::assert_snapshot!(get_stat(&work_dir, 1, 10), @r"
     1  | 10 +++
     一 | 10 +++
     2 files changed, 20 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Tracking 2 files
     [EOF]
     ");
 }
@@ -3776,6 +3870,10 @@ fn test_diff_stat_rounding() {
     mostly_removes.txt | 301 +--------------
     only_adds.txt      |  10 +
     3 files changed, 111 insertions(+), 301 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+      only_adds.txt
     [EOF]
     ");
 
@@ -3823,6 +3921,11 @@ fn test_diff_binary() {
         (binary)
     Added regular file binary_valid_utf8.png:
         (binary)
+    [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+      binary_added.png
+      binary_valid_utf8.png
     [EOF]
     ");
 
@@ -3893,6 +3996,10 @@ fn test_diff_stat_binary_and_text() {
     ...d_long_file_name.png | (binary)
     ...d_long_file_name.txt | 100 ++++++++++
     2 files changed, 100 insertions(+), 0 deletions(-)
+    [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+      text_with_elided_long_file_name.txt
     [EOF]
     ");